Preemptively fix warning

+++ benchmarks/output-exponentially-random-float/source_cstdio_g.cpp
@@ -7,7 +7,7 @@ int main() {
     std::minstd_rand gen;
     for (size_t i = 0; i < 500000; i++) {
         int32_t rng = gen();
-        float x = exp(rng / 46684427.);
+        float x = (float)exp(rng / 46684427.);
         if (rng & 1) {
             x = -x;
         }
